project/requirements.txt
fastapi
uvicorn
asyncpg
tortoise-orm